Moved time/date formatting to utils
